+++ elmboot/src/main/java/com/neusoft/elmboot/po/Orders.java
@@ -8,7 +8,7 @@ public class Orders {
 	private String userId;
 	private Integer businessId;
 	private String orderDate;
-	private Double orderTotal;
+	private Integer orderTotal;
 	private Integer daId; // 送货地址编号
 	private Integer orderState; // 订单状态（0：未支付； 1：已支付）
 
@@ -16,57 +16,75 @@ public class Orders {
 	private Business business;
 	// 一对多：订单明细
 	private List<OrderDetailet> list;
+
 	public Integer getOrderId() {
 		return orderId;
 	}
+
 	public void setOrderId(Integer orderId) {
 		this.orderId = orderId;
 	}
+
 	public String getUserId() {
 		return userId;
 	}
+
 	public void setUserId(String userId) {
 		this.userId = userId;
 	}
+
 	public Integer getBusinessId() {
 		return businessId;
 	}
+
 	public void setBusinessId(Integer businessId) {
 		this.businessId = businessId;
 	}
+
 	public String getOrderDate() {
 		return orderDate;
 	}
+
 	public void setOrderDate(String orderDate) {
 		this.orderDate = orderDate;
 	}
-	public Double getOrderTotal() {
+
+	public Integer getOrderTotal() {
 		return orderTotal;
 	}
-	public void setOrderTotal(Double orderTotal) {
+
+	public void setOrderTotal(Integer orderTotal) {
 		this.orderTotal = orderTotal;
 	}
+
 	public Integer getDaId() {
 		return daId;
 	}
+
 	public void setDaId(Integer daId) {
 		this.daId = daId;
 	}
+
 	public Integer getOrderState() {
 		return orderState;
 	}
+
 	public void setOrderState(Integer orderState) {
 		this.orderState = orderState;
 	}
+
 	public Business getBusiness() {
 		return business;
 	}
+
 	public void setBusiness(Business business) {
 		this.business = business;
 	}
+
 	public List<OrderDetailet> getList() {
 		return list;
 	}
+
 	public void setList(List<OrderDetailet> list) {
 		this.list = list;
 	}
